+++ anyblok/config.py
@@ -1074,7 +1074,7 @@ def add_schema(group):
     try:
         from graphviz.files import FORMATS
     except ImportError:
-        from graphviz.backend import FORMATS
+        from graphviz import FORMATS
 
     group.add_argument('--schema-format',
                        default='png', choices=tuple(FORMATS))