+++ doc/CHANGES.rst
@@ -17,7 +17,18 @@
 CHANGELOG
 =========
 
-1.3.0 (unreleased)
+1.4.0 (UNRELEASED)
+------------------
+
+* Fixed Graphviz dependencies #207
+* Fixed Mapper on column with **primary_key=False** #208
+
+1.3.1 (2021-08-25)
+------------------
+
+* Fixed sphinx pluggin 
+
+1.3.0 (2021-07-12)
 ------------------
 
 * Added deprecated name argument on the configuration to
@@ -27,6 +38,10 @@ CHANGELOG
   the sys arg
 * Added plugins to migrate DateTime to DateTime with MariaDB and MySQL
 * Added cache_ok attribute on Column Type
+* Fixed issue #195, the issue was due of a changement of the api of
+  the hybrid properties, Thanks @petrus-v for your help
+* Fixed issue #196, A setter on the deprecated attribut **registry**
+  do a setter on the attribute **anyblok**
 
 1.2.0 (2021-04-14)
 ------------------